requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.2
aiohttp==3.9.1
google-generativeai==0.3.2
httpx==0.26.0
aiosocks==0.2.6
